requirements.txt
google-generativeai
SpeechRecognition
pyttsx3
pyaudio
AppOpener
python-dotenv
